refactor: move utils impl block to dedicated module
