+++ honeycomb-core/src/cmap2/structure.rs
@@ -1,8 +1,12 @@
 //! mod doc
 
+// ------ IMPORTS
+
 use crate::cmap2::CMAP2_BETA;
 use crate::{AttrSparseVec, CoordsFloat, DartIdentifier, Vertex2};
 use std::collections::BTreeSet;
+
+// ------ CONTENT
 
 /// Main map object.
 ///
